feat: 게시글 상세 페이지 개선
