+++ src/pages/[username]/[slug]/index.tsx
@@ -1,10 +1,10 @@
 import { ReactElement } from "react"
 import { NextPageWithLayout } from "@pages/_app"
 import Layout from "@components/Layout"
-import Article from "@containers/Article"
+import ArticleDetail from "@containers/ArticleDetail"
 
 const ArticlePage: NextPageWithLayout = () => {
-  return <Article />
+  return <ArticleDetail />
 }
 
 ArticlePage.getLayout = (page: ReactElement) => {
